"Expanded Readme and improved scope related logging in router
The Readme file has been significantly expanded to detail the process of setting up routes, user middleware and the challenges encountered in handling the OpenAPI specification. In addition, the logging inside 'router.py' has been improved to specifically show all scopes required for a route instead of each individual scope due to an Authorize guard. This aims to provide a clearer understanding of the permissions needed for each route."