"Refactor user model and implement token permission check
Reorganized the project structure for improved readability, moving UserModel class to a separate module under domain/user directory. The UserModel class was also enhanced to include a permissions checking method. This was done to centralize the permission check logic within the UserModel itself, which makes it easier to manage and debug. Additionally, added a new CustomOperation class in the utilities directory to handle custom OpenAPI operations, and renamed dependencies and middleware files to make it consistent with the new domain-centric architecture."